extra3.o/requirements.txt
opencv-python
torch
torchvision
ultralytics
mysql-connector-python
easyocr
numpy
Flask
Flask-CORS
